+++ backend/src/index.ts
@@ -45,7 +45,6 @@ async function init() {
     if(feature("COMPILED")) Log.info("Running compiled executable");
     if(!feature("PRODUCTION")) Log.info("Using development build");
     if(flags.fakeData) Log.info("Using fake data for backend");
-    if(flags.noMAC) Log.info("MegaAntiCheat integration disabled");
     if(flags.noSteamApi) Log.info("Steam API usage disabled");
 
     await Promise.all([ Settings.init(), Values.init() ]);