+++ frontend/src/routes/setup/Integrations.svelte
@@ -5,11 +5,9 @@
     import { show } from "$lib/components/popups/ComponentPopup.svelte";
     import CircleQuestionMark from "@lucide/svelte/icons/circle-question-mark";
     import SteamKeyInfo from "$lib/components/settings/SteamKeyInfo.svelte";
-    import MACInfo from "$lib/components/settings/MACInfo.svelte";
     import SteamHistoryInfo from "$lib/components/settings/SteamHistoryInfo.svelte";
 
     watch(() => Settings.settings.steamApiKey, () => Settings.update("steamApiKey"));
-    watch(() => Settings.settings.masterbaseKey, () => Settings.update("masterbaseKey"));
     watch(() => Settings.settings.steamhistoryApiKey, () => Settings.update("steamhistoryApiKey"));
 </script>
 
@@ -22,15 +20,6 @@
 </h2>
 <input bind:value={Settings.settings.steamApiKey} type="password"
     placeholder="Steam API Key" class="w-full border-b border-gray-300 outline-none mt-1" />
-
-<h2 class="font-semibold text-lg mt-3">
-    MegaAntiCheat masterbase key
-    <button onclick={() => show(MACInfo)}>
-        <CircleQuestionMark size={16} class="text-gray-500" />
-    </button>
-</h2>
-<input bind:value={Settings.settings.masterbaseKey} type="password"
-    placeholder="MAC Api Key" class="w-full border-b border-gray-300 outline-none mt-1" />
 
 <h2 class="font-semibold text-lg mt-3">
     SteamHistory API Key