+++ .github/release-notes.md
@@ -2,3 +2,4 @@
 * tfview/respawned.cfg is automatically `exec`'d when respawning
 * Added dialog for when the user changes their casual selection externally
 * Missing avatars for users' friends are now fetched
+* Removed MegaAntiCheat integration